getLongestPrefixMatch() would enter a tight loop if the request URI didn't map into a context. Now, if a match can't be found it returns null and a 404 error page will be sent back to the client browser.
git-svn-id: https://svn.apache.org/repos/asf/tomcat@282921 13f79535-47bb-0310-9956-ffa450edef68

+++ container/tc3.x/trunk/src/share/org/apache/tomcat/util/PrefixMapper.java
@@ -227,6 +227,9 @@ public class PrefixMapper {
 	    container = myMap.prefixMappedServlets.get(s);
 	    
 	    if (container == null) {
+		// if empty string didn't map, time to give up
+		if ( s.length() == 0 )
+                    break;
 		s=FileUtil.removeLast( s );
 	    }  else {
 		if( myMap.mapCacheEnabled ) {
